refactor: package config flow and services helpers
Move 16 root-level config_flow_*.py helper modules into
custom_components/thessla_green_modbus/config_flow/ and 13
services_*.py helper modules into
custom_components/thessla_green_modbus/services/.

config_flow/ package (16 files):
  config_flow.py            → config_flow/__init__.py  (HA entrypoint)
  config_flow_confirm.py    → config_flow/confirm.py
  config_flow_device_validation.py → config_flow/device_validation.py
  config_flow_entry.py      → config_flow/entry.py
  config_flow_errors.py     → config_flow/errors.py
  config_flow_network.py    → config_flow/network.py
  config_flow_options.py    → config_flow/options.py
  config_flow_options_form.py → config_flow/options_form.py
  config_flow_payloads.py   → config_flow/payloads.py
  config_flow_reauth.py     → config_flow/reauth.py
  config_flow_reauth_confirm.py → config_flow/reauth_confirm.py
  config_flow_runtime.py    → config_flow/runtime.py
  config_flow_schema.py     → config_flow/schema.py
  config_flow_steps.py      → config_flow/steps.py
  config_flow_user_submit.py → config_flow/user_submit.py
  config_flow_validation.py → config_flow/validation.py

services/ package (13 files):
  services.py               → services/__init__.py
  services_dispatch.py      → services/dispatch.py
  services_handler_deps.py  → services/handler_deps.py
  services_handlers_data.py → services/handlers_data.py
  services_handlers_logging.py → services/handlers_logging.py
  services_handlers_maintenance.py → services/handlers_maintenance.py
  services_handlers_mode.py → services/handlers_mode.py
  services_handlers_parameters.py → services/handlers_parameters.py
  services_handlers_schedule.py → services/handlers_schedule.py
  services_helpers.py       → services/helpers.py
  services_schema.py        → services/schema.py
  services_targets.py       → services/targets.py
  services_validation.py    → services/validation.py

All sibling imports (from .config_flow_X / from .services_X) updated
to short-form relative imports (from .X). All parent-package imports
(from .const, .errors, .modbus_exceptions, etc.) updated to double-dot
form (from ..X). Import order re-sorted by ruff where needed.

Test files updated to reflect new module paths:
  test_api_contracts.py, test_config_flow_helpers.py,
  test_config_flow_options.py, test_config_flow_runtime_validation.py,
  test_services_dispatch_validation.py,
  test_services_handlers_maintenance.py,
  test_services_handlers_parameters_registration.py

HA entrypoint unchanged: config_flow/__init__.py is resolved by Python
at the same dotted path custom_components.thessla_green_modbus.config_flow.
Service IDs, register names, entity IDs, and runtime behavior unchanged.
