+++ custom_components/thessla_green_modbus/services/schema.py
